# Add flash attention support for FP8 Mistral