@regisss 2024-08-07
Closing as it was integrated into #1163.